alpha 1.0

+++ routes/web.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\ReportController;
 use App\Http\Controllers\SaleController;
 use App\Http\Controllers\SettingsController;
 use App\Http\Controllers\StaffController;
+use App\Http\Controllers\StoreController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -17,7 +18,7 @@ Route::get('/', function () {
 });
 
 Route::get('/login', [StaffAuthController::class, 'showLoginForm'])->name('login');
-Route::post('/login', [StaffAuthController::class, 'login']);
+Route::post('/login', [StaffAuthController::class, 'login'])->middleware('throttle:5,1');
 Route::post('/logout', [StaffAuthController::class, 'logout'])->name('logout');
 
 Route::middleware(['auth:staff'])->group(function () {
@@ -59,10 +60,10 @@ Route::middleware(['auth:staff'])->group(function () {
     Route::get('/staff', [StaffController::class, 'index'])->name('staff.index');
     Route::get('/staff/{staff}', [StaffController::class, 'show'])->name('staff.show');
     Route::get('/staff/{staff}/edit', [StaffController::class, 'edit'])->name('staff.edit');
-    Route::post('/staff', [StaffController::class, 'store'])->name('staff.store');
-    Route::put('/staff/{staff}', [StaffController::class, 'update'])->name('staff.update');
-    Route::post('/staff/{staff}/password', [StaffController::class, 'changePassword'])->name('staff.password');
-    Route::delete('/staff/{staff}', [StaffController::class, 'destroy'])->name('staff.destroy');
+    Route::post('/staff', [StaffController::class, 'store'])->name('staff.store')->middleware('role:Administrador');
+    Route::put('/staff/{staff}', [StaffController::class, 'update'])->name('staff.update')->middleware('role:Administrador');
+    Route::post('/staff/{staff}/password', [StaffController::class, 'changePassword'])->name('staff.password')->middleware('role:Administrador');
+    Route::delete('/staff/{staff}', [StaffController::class, 'destroy'])->name('staff.destroy')->middleware('role:Administrador');
 
     Route::get('/reports', [ReportController::class, 'index'])->name('reports.index');
     Route::get('/reports/direct/{date}', [ReportController::class, 'directSales'])->name('reports.direct');
@@ -84,6 +85,40 @@ Route::middleware(['auth:staff'])->group(function () {
 
     Route::get('/settings', [SettingsController::class, 'index'])->name('settings.index');
     Route::put('/settings', [SettingsController::class, 'update'])->name('settings.update');
+
+    Route::group(['middleware' => 'role:Administrador,Vendedor'], function () {
+        Route::get('/store', [StoreController::class, 'index'])->name('store.index');
+        Route::get('/store/productos', [StoreController::class, 'products'])->name('store.products');
+        Route::get('/store/productos/crear', [StoreController::class, 'productCreate'])->name('store.products.create');
+        Route::post('/store/productos', [StoreController::class, 'productStore'])->name('store.products.store');
+        Route::get('/store/productos/ventas', [StoreController::class, 'productSales'])->name('store.products.sales');
+        Route::post('/store/productos/venta', [StoreController::class, 'productProcessSale'])->name('store.products.sale.process');
+        Route::post('/store/productos/add-stock', [StoreController::class, 'productAddStock'])->name('store.products.add-stock');
+        Route::get('/store/productos/reporte', [StoreController::class, 'productReport'])->name('store.products.report');
+        Route::get('/store/productos/reporte/imagenes', [StoreController::class, 'productImagesReport'])->name('store.products.report.images');
+        Route::get('/store/productos/reporte/ventas', [StoreController::class, 'productSalesReport'])->name('store.products.report.sales');
+        Route::post('/store/productos/tipos', [StoreController::class, 'productTypeStore'])->name('store.products.types.store');
+        Route::post('/store/productos/tipos/{storeProductType}/delete', [StoreController::class, 'productTypeDestroy'])->name('store.products.types.destroy');
+        Route::get('/store/productos/{storeProduct}', [StoreController::class, 'productShow'])->name('store.products.show');
+        Route::get('/store/productos/{storeProduct}/editar', [StoreController::class, 'productEdit'])->name('store.products.edit');
+        Route::put('/store/productos/{storeProduct}', [StoreController::class, 'productUpdate'])->name('store.products.update');
+        Route::delete('/store/productos/{storeProduct}', [StoreController::class, 'productDestroy'])->name('store.products.destroy');
+        Route::get('/store/regalos', [StoreController::class, 'gifts'])->name('store.gifts');
+        Route::get('/store/regalos/crear', [StoreController::class, 'giftCreate'])->name('store.gifts.create');
+        Route::post('/store/regalos', [StoreController::class, 'giftStore'])->name('store.gifts.store');
+        Route::get('/store/regalos/ventas', [StoreController::class, 'giftSales'])->name('store.gifts.sales');
+        Route::post('/store/regalos/venta', [StoreController::class, 'giftProcessSale'])->name('store.gifts.sale.process');
+        Route::post('/store/regalos/add-stock', [StoreController::class, 'giftAddStock'])->name('store.gifts.add-stock');
+        Route::get('/store/regalos/reporte', [StoreController::class, 'giftReport'])->name('store.gifts.report');
+        Route::get('/store/regalos/reporte/imagenes', [StoreController::class, 'giftImagesReport'])->name('store.gifts.report.images');
+        Route::get('/store/regalos/reporte/ventas', [StoreController::class, 'giftSalesReport'])->name('store.gifts.report.sales');
+        Route::post('/store/regalos/tipos', [StoreController::class, 'giftTypeStore'])->name('store.gifts.types.store');
+        Route::post('/store/regalos/tipos/{storeGiftType}/delete', [StoreController::class, 'giftTypeDestroy'])->name('store.gifts.types.destroy');
+        Route::get('/store/regalos/{storeGift}', [StoreController::class, 'giftShow'])->name('store.gifts.show');
+        Route::get('/store/regalos/{storeGift}/editar', [StoreController::class, 'giftEdit'])->name('store.gifts.edit');
+        Route::put('/store/regalos/{storeGift}', [StoreController::class, 'giftUpdate'])->name('store.gifts.update');
+        Route::delete('/store/regalos/{storeGift}', [StoreController::class, 'giftDestroy'])->name('store.gifts.destroy');
+    });
 });
 
 Route::get('/home', function () {
